+++ CHANGELOG.md
@@ -61,4 +61,5 @@
 
 # DEVEL version
 
-  - Sysrepo version bumped from 0.7.7 to 660552222ee6376efa560d9bcc7b832886ff460a 
+  - Sysrepo version bumped from 0.7.7 to 660552222ee6376efa560d9bcc7b832886ff460a
+  - Implemented dump_xpaths() to sysrepo data abstraction layer.